# docs/api/openapi.yml
openapi: 3.0.0
info:
    title: WildDuck API
    description: WildDuck API docs
    version: 1.0.0
    contact:
        url: 'https://github.com/nodemailer/wildduck'
servers:
    - url: 'https://api.wildduck.email'
...
paths:
...
    '/users/{user}/storage/{file}':
        delete:
            tags:
                - Storage
            summary: Delete a File
            operationId: deleteFile
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
        get:
            tags:
                - Storage
            summary: Download File
            description: This method returns stored file contents in binary form
            operationId: getFile
            responses:
                '200':
                    description: Success
                    content:
                        application/octet-stream:
                            schema:
                                type: string
                                description: Success
                                format: binary
        parameters:
            - name: user
              in: path
              description: ID of the User
              required: true
              schema:
                  type: string
            - name: file
              in: path
              description: ID of the File
              required: true
              schema:
                  type: string
    '/users/{user}/storage':
        get:
            tags:
                - Storage
            summary: List stored files
            operationId: getFiles
            parameters:
                - name: query
                  in: query
                  description: Partial match of a filename
                  schema:
                      type: string
                - name: limit
                  in: query
                  description: How many records to return
                  schema:
                      type: number
                - name: page
                  in: query
                  description: 'Current page number. Informational only, page numbers start from 1'
                  schema:
                      type: number
                - name: next
                  in: query
                  description: 'Cursor value for next page, retrieved from nextCursor response value'
                  schema:
                      type: number
                - name: previous
                  in: query
                  description: 'Cursor value for previous page, retrieved from previousCursor response value'
                  schema:
                      type: number
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetFilesResponse'
        post:
            tags:
                - Storage
            summary: Upload File
            description: This method allows to upload an attachment to be linked from a draft
            operationId: uploadFile
            requestBody:
                content:
                    multipart/form-data:
                        encoding: {}
                        schema:
                            required:
                                - body
                            type: object
                            properties:
                                body:
                                    type: string
                                    format: binary
                required: false
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UploadFileResponse'
        parameters:
            - name: user
              in: path
              description: ID of the User
              required: true
              schema:
                  type: string
...
components:
    schemas:
...
        SuccessResponse:
            required:
                - success
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
...
        UploadFileResponse:
            required:
                - success
                - id
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                id:
                    type: object
                    description: File ID
...
        GetFilesResponse:
            required:
                - success
                - total
                - page
                - previousCursor
                - nextCursor
                - results
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                total:
                    type: number
                    description: How many results were found
                page:
                    type: number
                    description: Current page number. Derived from page query argument
                previousCursor:
                    type: string
                    description: Either a cursor string or false if there are not any previous results
                nextCursor:
                    type: string
                    description: Either a cursor string or false if there are not any next results
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetFilesResult'
                    description: File listing
...
        GetFilesResult:
            required:
                - id
                - filename
                - contentType
                - size
            type: object
            properties:
                id:
                    type: string
                    description: ID of the File
                filename:
                    type: string
                    description: Filename
                contentType:
                    type: string
                    description: Content-Type of the file
                size:
                    type: number
                    description: File size
...
    securitySchemes:
        AccessTokenAuth:
            name: X-Access-Token
            type: apiKey
            in: header
            description: |-
                If authentication is enabled in the WildDuck configuration, you will need to supply an access token in the `X-Access-Token` header.

                ```json
                {
                  "X-Access-Token": "59fc66a03e54454869460e45"
                }
                ```
security:
    - AccessTokenAuth: []